Don't dump into files

+++ hywiki/pipelines.py
@@ -7,10 +7,14 @@ class HywikiJSONPipeline(object):
     self.num_items = 0
     self.items = defaultdict(int)
     self.denominator = 10
+    self.already_dumped_counts = set()
 
   def dump_json(self, prefix=''):
+    if self.num_items in self.already_dumped_counts:
+      return
     with open('%sitems-%d.json' % (prefix, self.num_items), 'w') as outfile:
       json.dump(self.items, outfile, indent=2)
+      self.already_dumped_counts.add(self.num_items)
 
   def process_item(self, item, spider):
     self.items[item['text']] += 1
